Add SharedComponent helper interface for components like ADS1x15
For components that do multiplexing and are used by multiple sensor
components, such as the ADS1115Component / ADS1115Sensor, add a helper
class that allows the users of those sensors to request continuous
periods of usage time of the parent component.  If this new
request_time() interface is used, measurements made on one child sensor
will not interrupt measurements happening on another child sensor
sharing the same parent component.

CT clamp is where this is going to be useful.